+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Bumps Instabug iOS SDK to v11.3.0
 * Adds BugReporting.setVideoRecordingFloatingButtonPosition API
+* Adds some missing string keys and their mappings on iOS and Android
+* Adds missing mapping for portuguesePortugal locale on iOS
+* Fixes an issue with some string keys not working on Android
 
 ## 11.2.0 (2022-09-08)